+++ ExpressionParser.h
@@ -6,7 +6,7 @@
 class ExpressionParser
 {
 	mu::Parser parser;
-	static std::map<std::string, double> variables;
+	static std::map<std::string, float> variables;
 	static std::map<std::string, std::string> expressions;
 	void loadVariables();
 	std::string myExpression;
@@ -14,12 +14,12 @@ public:
 	ExpressionParser();
 	~ExpressionParser();
 	
-	double evaluate();
+	float evaluate();
 
 	void setString(std::string);
 	static void updateVariable(std::string name, std::string expr);
 	static void updateExpression(std::string name, std::string expr);
-	void addLocalVariable(std::string name, double* location);
+	void addLocalVariable(std::string name, float* location);
 	
 };
